+++ SolastaUnfinishedBusiness/Models/SrdAndHouseRulesContext.cs
@@ -7,17 +7,12 @@ using SolastaUnfinishedBusiness.Builders;
 using SolastaUnfinishedBusiness.Builders.Features;
 using SolastaUnfinishedBusiness.CustomBehaviors;
 using SolastaUnfinishedBusiness.CustomInterfaces;
-using SolastaUnfinishedBusiness.CustomUI;
-using SolastaUnfinishedBusiness.CustomValidators;
-using SolastaUnfinishedBusiness.Properties;
-using SolastaUnfinishedBusiness.Subclasses;
 using static RuleDefinitions;
 using static SolastaUnfinishedBusiness.Api.DatabaseHelper;
 using static SolastaUnfinishedBusiness.Api.DatabaseHelper.ConditionDefinitions;
 using static SolastaUnfinishedBusiness.Api.DatabaseHelper.FeatureDefinitionActionAffinitys;
 using static SolastaUnfinishedBusiness.Api.DatabaseHelper.SpellDefinitions;
 using static SolastaUnfinishedBusiness.Api.DatabaseHelper.MonsterDefinitions;
-using static SolastaUnfinishedBusiness.Api.DatabaseHelper.FeatureDefinitionCastSpells;
 using static SolastaUnfinishedBusiness.Api.DatabaseHelper.FeatureDefinitionSenses;
 using static SolastaUnfinishedBusiness.Api.DatabaseHelper.CharacterClassDefinitions;
 using static SolastaUnfinishedBusiness.Api.DatabaseHelper.ItemDefinitions;
@@ -64,89 +59,41 @@ internal static class SrdAndHouseRulesContext
 
     private static readonly string[] LargeWildshapeForms =
     {
-        "WildShapeAirElemental", "WildShapeFireElemental", "WildShapeEarthElemental", "WildShapeWaterElemental",
-        "WildShapeApe", "WildShapeTundraTiger"
+        "WildShapeAirElemental", //
+        "WildShapeFireElemental", //
+        "WildShapeEarthElemental", //
+        "WildShapeWaterElemental", //
+        "WildShapeApe", //
+        "WildShapeTundraTiger" //
     };
 
     private static readonly Dictionary<string, TagsDefinitions.Criticity> Tags = new();
 
     private static SpellDefinition ConjureElementalInvisibleStalker { get; set; }
 
-    internal static void Load()
+    internal static void LateLoad()
     {
-        //BUGFIX: these null shouldn't be there as it breaks Bard Magical Secrets
-        foreach (var spells in SpellListDefinitions.SpellListAllSpells.SpellsByLevel.Select(x => x.Spells))
-        {
-            spells.RemoveAll(x => x == null);
-        }
-
-        //BUGFIX: add a sprite reference to Resurrection
-        Resurrection.GuiPresentation.spriteReference =
-            Sprites.GetSprite("Resurrection", Resources.Resurrection, 128, 128);
-
-        //BUGFIX: this official condition doesn't have sprites or description
-        ConditionDefinitions.ConditionConjuredItemLink.silentWhenAdded = true;
-        ConditionDefinitions.ConditionConjuredItemLink.silentWhenRemoved = true;
-        ConditionDefinitions.ConditionConjuredItemLink.GuiPresentation.hidden = true;
-
-        //BUGFIX: add an effect to Counterspell
-        Counterspell.EffectDescription.effectParticleParameters =
-            DreadfulOmen.EffectDescription.effectParticleParameters;
-
-        // BUGFIX: Chill Touch and Ray of Frost should have not saving throw
-        ChillTouch.EffectDescription.EffectForms[0].savingThrowAffinity = EffectSavingThrowType.None;
-        RayOfFrost.EffectDescription.EffectForms[0].savingThrowAffinity = EffectSavingThrowType.None;
-
-        //BEHAVIOR: Allow Duelist higher level feature to interact correctly with Uncanny Dodge
-        static IsCharacterValidHandler IsActionAffinityUncannyDodgeValid(params string[] conditions)
-        {
-            // this allows Reflexive Party to trigger without Uncanny Dodge which can be triggered after that
-            return character => character.GetSubclassLevel(Rogue, RoguishDuelist.Name) < 13 ||
-                                conditions.Any(character.HasConditionOfType);
-        }
-
-        ActionAffinityUncannyDodge.SetCustomSubFeatures(new ValidatorsDefinitionApplication(
-            IsActionAffinityUncannyDodgeValid(RoguishDuelist.ConditionReflexiveParry)));
-
         //SETTING: modify normal vision range
         SenseNormalVision.senseRange = Main.Settings.IncreaseSenseNormalVision;
 
+        AddBleedingToRestoration();
         AllowTargetingSelectionWhenCastingChainLightningSpell();
         ApplyConditionBlindedShouldNotAllowOpportunityAttack();
         ApplySrdWeightToFoodRations();
         BuildConjureElementalInvisibleStalker();
         LoadAfterRestIdentify();
-        UngroupWildshapeAttacks();
-    }
-
-    internal static void LateLoad()
-    {
-        FixDivineSmiteRestrictions();
-        FixDivineSmiteDiceAndBrandingSmiteNumberWhenUsingHighLevelSlots();
-        FixMeleeHitEffectsRange();
-        FixMountaineerBonusShoveRestrictions();
-        FixRecklessAttackForReachWeapons();
-        FixStunningStrikeForAnyMonkWeapon();
-        SpellsMinorFixes();
-        AddBleedingToRestoration();
-        SwitchFilterOnHideousLaughter();
-        SwitchRecurringEffectOnEntangle();
-        UseCubeOnSleetStorm();
-        SwitchEldritchBlastRange();
-        UseHeightOneCylinderEffect();
-        SwitchUniversalSylvanArmorAndLightbringer();
-        SwitchDruidAllowMetalArmor();
-        SwitchMagicStaffFoci();
-        SwitchEnableUpcastConjureElementalAndFey();
-        SwitchFullyControlConjurations();
-        SwitchMakeLargeWildshapeFormsMedium();
         SwitchAllowClubsToBeThrown();
-        FixMartialArtsProgression();
-        FixTwinnedMetamagic();
-        FixAttackBuffsAffectingSpellDamage();
-        FixMissingWildShapeTagOnSomeForms();
-        MakeGorillaWildShapeRocksUnlimited();
-        AddCustomWeaponValidatorToFightingStyleArchery();
+        SwitchDruidAllowMetalArmor();
+        SwitchEldritchBlastRange();
+        SwitchEnableUpcastConjureElementalAndFey();
+        SwitchFilterOnHideousLaughter();
+        SwitchFullyControlConjurations();
+        SwitchMagicStaffFoci();
+        SwitchMakeLargeWildshapeFormsMedium();
+        SwitchRecurringEffectOnEntangle();
+        SwitchUniversalSylvanArmorAndLightbringer();
+        UseCubeOnSleetStorm();
+        UseHeightOneCylinderEffect();
     }
 
     internal static void SwitchUniversalSylvanArmorAndLightbringer()
@@ -199,80 +146,6 @@ internal static class SrdAndHouseRulesContext
             item.IsFocusItem = true;
             item.FocusItemDescription.focusType = EquipmentDefinitions.FocusType.Arcane;
         }
-    }
-
-    /**
-     * Makes Divine Smite trigger only from melee attacks.
-     * This wasn't relevant until we changed how SpendSpellSlot trigger works.
-     */
-    private static void FixDivineSmiteRestrictions()
-    {
-        FeatureDefinitionAdditionalDamages.AdditionalDamagePaladinDivineSmite.attackModeOnly = true;
-        FeatureDefinitionAdditionalDamages.AdditionalDamagePaladinDivineSmite.requiredProperty =
-            RestrictedContextRequiredProperty.MeleeWeapon;
-    }
-
-    /**
-     * Makes Divine Smite use correct number of dice when spending slot level 5+.
-     * Base game has config only up to level 4 slots, which leads to it using 1 die if level 5+ slot is spent.
-     */
-    private static void FixDivineSmiteDiceAndBrandingSmiteNumberWhenUsingHighLevelSlots()
-    {
-        FeatureDefinitionAdditionalDamages.AdditionalDamagePaladinDivineSmite.diceByRankTable =
-            DiceByRankBuilder.BuildDiceByRankTable(2);
-
-        FeatureDefinitionAdditionalDamages.AdditionalDamageBrandingSmite.diceByRankTable =
-            DiceByRankBuilder.BuildDiceByRankTable(2);
-    }
-
-    /**
-     * Ensures any spell or power effect in game that uses MeleeHit has a correct range of 1.
-     * Otherwise our AttackEvaluationParams.FillForMagicReachAttack will use incorrect data.
-     */
-    private static void FixMeleeHitEffectsRange()
-    {
-        foreach (var effectDescription in DatabaseRepository.GetDatabase<SpellDefinition>()
-                     .Select(x => x.EffectDescription)
-                     .Where(x => x.rangeType == RangeType.MeleeHit))
-        {
-            effectDescription.rangeParameter = 1;
-        }
-
-        foreach (var effectDescription in DatabaseRepository.GetDatabase<FeatureDefinitionPower>()
-                     .Select(x => x.EffectDescription)
-                     .Where(x => x.rangeType == RangeType.MeleeHit))
-        {
-            effectDescription.rangeParameter = 1;
-        }
-    }
-
-    /**
-     * Makes Mountaineer's `Shield Push` bonus shove work only with shield equipped.
-     * This wasn't relevant until we removed forced shield check in the `GameLocationCharacter.GetActionStatus`.
-     */
-    private static void FixMountaineerBonusShoveRestrictions()
-    {
-        ActionAffinityMountaineerShieldCharge
-            .SetCustomSubFeatures(new ValidatorsDefinitionApplication(ValidatorsCharacter.HasShield));
-    }
-
-    /**
-     * Makes `Reckless` context check if main hand weapon is melee, instead of if character is next to target.
-     * Required for it to work on reach weapons.
-     */
-    private static void FixRecklessAttackForReachWeapons()
-    {
-        FeatureDefinitionCombatAffinitys.CombatAffinityReckless
-            .situationalContext = (SituationalContext)ExtraSituationalContext.MainWeaponIsMeleeOrUnarmed;
-    }
-
-    /**
-     * Makes `Stunning Strike` context check if any monk weapon instead on OnAttackMeleeHitAuto
-     * Required for it to work with monk weapon specialization and/or way of distant hand.
-     */
-    private static void FixStunningStrikeForAnyMonkWeapon()
-    {
-        FeatureDefinitionPowers.PowerMonkStunningStrike.activationTime = ActivationTime.OnAttackHitAuto;
     }
 
     internal static void ApplyConditionBlindedShouldNotAllowOpportunityAttack()
@@ -356,28 +229,6 @@ internal static class SrdAndHouseRulesContext
         {
             Entangle.effectDescription.recurrentEffect =
                 RecurrentEffect.OnActivation | RecurrentEffect.OnTurnEnd | RecurrentEffect.OnEnter;
-        }
-    }
-
-    private static void SpellsMinorFixes()
-    {
-        // Shows Concentration tag in UI
-        BladeBarrier.requiresConcentration = true;
-
-        //
-        // BUGFIX: spells durations
-        //
-
-        // Stops upcasting assigning non-SRD durations
-        var spells = new IMagicEffect[]
-        {
-            ProtectionFromEnergy, ProtectionFromEnergyAcid, ProtectionFromEnergyCold, ProtectionFromEnergyFire,
-            ProtectionFromEnergyLightning, ProtectionFromEnergyThunder, ProtectionFromPoison
-        };
-
-        foreach (var spell in spells)
-        {
-            spell.EffectDescription.EffectAdvancement.alteredDuration = AdvancementDuration.None;
         }
     }
 
@@ -570,96 +421,6 @@ internal static class SrdAndHouseRulesContext
         }
     }
 
-    private static void FixMartialArtsProgression()
-    {
-        //Fixes die progression of Monk's Martial Arts to use Monk level, not character level
-        var provider = new RankByClassLevel(Monk);
-        var features = new List<FeatureDefinition>
-        {
-            FeatureDefinitionAttackModifiers.AttackModifierMonkMartialArtsImprovedDamage,
-            FeatureDefinitionAttackModifiers.AttackModifierMonkMartialArtsUnarmedStrikeBonus,
-            FeatureDefinitionAttackModifiers.AttackModifierMonkFlurryOfBlowsUnarmedStrikeBonus,
-            FeatureDefinitionAttackModifiers.AttackModifierMonkFlurryOfBlowsUnarmedStrikeBonusFreedom
-        };
-
-        foreach (var feature in features)
-        {
-            feature.AddCustomSubFeatures(provider);
-        }
-    }
-
-    private static void FixTwinnedMetamagic()
-    {
-        //BUGFIX: fix vanilla twinned spells offering not accounting for target parameter progression
-        MetamagicOptionDefinitions.MetamagicTwinnedSpell.AddCustomSubFeatures(new MetamagicApplicationValidator(
-            (RulesetCharacter _, RulesetEffectSpell spell, MetamagicOptionDefinition _, ref bool result,
-                ref string failure) =>
-            {
-                var effectDescription = spell.SpellDefinition.effectDescription;
-
-                if (effectDescription.TargetType is not (TargetType.Individuals or TargetType.IndividualsUnique)
-                    || spell.ComputeTargetParameter() == 1)
-                {
-                    return;
-                }
-
-                failure = FailureFlagInvalidSingleTarget;
-                result = false;
-            }));
-    }
-
-    private static void FixAttackBuffsAffectingSpellDamage()
-    {
-        //BUGFIX: fix Branding Smite applying bonus damage to spells
-        FeatureDefinitionAdditionalDamages.AdditionalDamageBrandingSmite
-            .AddCustomSubFeatures(ValidatorsRestrictedContext.WeaponAttack);
-
-        //BUGFIX: fix Divine Favor applying bonus damage to spells
-        FeatureDefinitionAdditionalDamages.AdditionalDamageDivineFavor
-            .AddCustomSubFeatures(ValidatorsRestrictedContext.WeaponAttack);
-    }
-
-    private static void FixMissingWildShapeTagOnSomeForms()
-    {
-        //BUGFIX: fix some Wild Shape forms missing proper tag, making wild shape action button visible while wild-shaped
-        var wildShape = FeatureDefinitionPowers.PowerDruidWildShape;
-        foreach (var option in wildShape.EffectDescription.FindFirstShapeChangeForm().ShapeOptions)
-        {
-            option.substituteMonster.CreatureTags.TryAdd(TagsDefinitions.CreatureTagWildShape);
-        }
-    }
-
-    private static void MakeGorillaWildShapeRocksUnlimited()
-    {
-        //CHANGE: makes Wildshape Gorilla form having unlimited rock toss attacks 
-        MonsterAttackDefinitions.Attack_Wildshape_Ape_Toss_Rock.limitedUse = false;
-        MonsterAttackDefinitions.Attack_Wildshape_Ape_Toss_Rock.maxUses = -1;
-    }
-
-    private static void UngroupWildshapeAttacks()
-    {
-        WildShapeApe.groupAttacks = false;
-        WildshapeBlackBear.groupAttacks = false;
-        WildShapeBrownBear.groupAttacks = false;
-        WildshapeDeepSpider.groupAttacks = false;
-        WildShapeGiant_Eagle.groupAttacks = false;
-    }
-
-    // allow darts, lightning launcher or hand crossbows benefit from Archery Fighting Style
-    private static void AddCustomWeaponValidatorToFightingStyleArchery()
-    {
-        FeatureDefinitionAttackModifiers.AttackModifierFightingStyleArchery.SetCustomSubFeatures(
-            new RestrictedContextValidator((_, _, _, item, _, _, _) => (OperationType.Set,
-                ValidatorsWeapon.IsWeaponType(item,
-                    CustomWeaponsContext.HandXbowWeaponType,
-                    CustomWeaponsContext.LightningLauncherType,
-                    WeaponTypeDefinitions.LongbowType,
-                    WeaponTypeDefinitions.ShortbowType,
-                    WeaponTypeDefinitions.HeavyCrossbowType,
-                    WeaponTypeDefinitions.LightCrossbowType,
-                    WeaponTypeDefinitions.DartType))));
-    }
-
     internal static void SwitchEnableUpcastConjureElementalAndFey()
     {
         if (!Main.Settings.EnableUpcastConjureElementalAndFey)
@@ -671,8 +432,8 @@ internal static class SrdAndHouseRulesContext
 
         ConfigureAdvancement(ConjureFey);
         ConfigureAdvancement(ConjureElemental);
-        ConjureElemental.SubspellsList.Add(ConjureElementalInvisibleStalker);
         ConfigureAdvancement(ConjureMinorElementals);
+        ConjureElemental.SubspellsList.Add(ConjureElementalInvisibleStalker);
 
         // Set advancement at spell level, not sub-spell
         static void ConfigureAdvancement([NotNull] IMagicEffect spell)
